Add files via upload
Austo save was failing when the filament had already unloaded from the CFS.  The regular save was working when pressed during the print.  I now save off the last used filament slot and use that as the save slot locatio.  Must be a single color Calibration print to work correctly, but this is as intended.